+++ buildingsync/views/main.py
@@ -1,0 +1,6 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views.generic import TemplateView
+
+
+class IndexView(LoginRequiredMixin, TemplateView):
+    template_name = 'buildingsync/index.html'
